+++ src/components/notices/NoticeListDropDownMenu.tsx
@@ -1,31 +1,77 @@
-import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+import { Check, ChevronsUpDown } from "lucide-react";
+import * as React from "react";
 
-export default function NoticeDropdownMenu() {
+import { Button } from "@/components/ui/button";
+import { Command, CommandGroup, CommandItem } from "@/components/ui/command";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
+import { cn } from "@/lib/utils";
+
+const frameworks = [
+  {
+    value: "test1",
+    label: "마감임박순",
+  },
+  {
+    value: "test2",
+    label: "시급많은순",
+  },
+  {
+    value: "test3",
+    label: "시급적은순",
+  },
+  {
+    value: "test4",
+    label: "가나다순",
+  },
+];
+
+export default function NoticeListDropdownMenu() {
+  const [open, setOpen] = React.useState(false);
+  const [value, setValue] = React.useState("");
+
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
+    <Popover open={open} onOpenChange={setOpen}>
+      <PopoverTrigger asChild>
         <Button
-          className="w-[auto] justify-between bg-gray-10 text-[1.2rem] font-bold "
           variant="outline"
+          role="combobox"
+          aria-expanded={open}
+          className="w-[auto] justify-between bg-gray-10 text-[1.2rem] font-bold "
         >
-          Open
+          {value
+            ? frameworks.find((framework) => framework.value === value)?.label
+            : "마감임박순"}
+          <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
         </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent className="w-56">
-        <DropdownMenuGroup>
-          <DropdownMenuItem>마감임박순</DropdownMenuItem>
-          <DropdownMenuItem>시급많은순</DropdownMenuItem>
-          <DropdownMenuItem>시급적은순</DropdownMenuItem>
-          <DropdownMenuItem>가나다순</DropdownMenuItem>
-        </DropdownMenuGroup>
-      </DropdownMenuContent>
-    </DropdownMenu>
+      </PopoverTrigger>
+      <PopoverContent className="w-[200px] p-0">
+        <Command>
+          <CommandGroup>
+            {frameworks.map((framework) => (
+              <CommandItem
+                key={framework.value}
+                value={framework.value}
+                onSelect={(currentValue: string) => {
+                  setValue(currentValue === value ? "" : currentValue);
+                  setOpen(false);
+                }}
+              >
+                <Check
+                  className={cn(
+                    "mr-2 h-4 w-4",
+                    value === framework.value ? "opacity-100" : "opacity-0",
+                  )}
+                />
+                {framework.label}
+              </CommandItem>
+            ))}
+          </CommandGroup>
+        </Command>
+      </PopoverContent>
+    </Popover>
   );
 }
